module03 finish

+++ module_01/ex05/Brain.cpp
@@ -6,7 +6,7 @@
 /*   By: jeldora <jeldora@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/24 02:23:49 by jeldora           #+#    #+#             */
-/*   Updated: 2020/10/24 23:36:43 by jeldora          ###   ########.fr       */
+/*   Updated: 2020/10/28 02:31:42 by jeldora          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 Brain::Brain()
 {
-	srand(static_cast<unsigned int>(time(0)));
+	srand((time(0)));
 	iq = rand() % 100;
 	if (iq > 200)
 		iq = 200;
